+++ ICD.Connect.Settings/Core/AbstractCoreSettings.cs
@@ -1,7 +1,0 @@
-﻿namespace ICD.Connect.Settings.Core
-{
-	public abstract class AbstractCoreSettings : AbstractSettings, ICoreSettings
-	{
-		public abstract SettingsCollection OriginatorSettings { get; }
-	}
-}