+++ ICD.Connect.Settings/Originators/IOriginator.cs
@@ -6,10 +6,9 @@ using ICD.Common.Utils;
 using ICD.Common.Utils.Services.Logging;
 using ICD.Connect.API.Attributes;
 using ICD.Connect.API.Nodes;
-using ICD.Connect.Settings.Core;
 using ICD.Connect.Settings.Proxies;
 
-namespace ICD.Connect.Settings
+namespace ICD.Connect.Settings.Originators
 {
 	/// <summary>
 	/// IOriginator represents an object that has settings.