+++ ICD.Connect.Settings/Originators/Simpl/AbstractSimplOriginator.cs
@@ -1,6 +1,6 @@
-﻿using System;
+using System;
 
-namespace ICD.Connect.Settings.Simpl
+namespace ICD.Connect.Settings.Originators.Simpl
 {
 	public abstract class AbstractSimplOriginator<TSettings> : AbstractOriginator<TSettings>, ISimplOriginator
 		where TSettings : ISimplOriginatorSettings, new()